% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextLabels.R
\name{calculateLabelHeightsAndWidths}
\alias{calculateLabelHeightsAndWidths}
\title{Calculate the heights and widths of the labels in the current plotting window}
\usage{
calculateLabelHeightsAndWidths(labels, cex, polygonAdded)
}
\arguments{
\item{labels}{A vector of the labels}

\item{cex}{The number used to scale the size of the label and therefore its height and width}

\item{polygonAdded}{A logical (TRUE/FALSE) variable to indicate whether polygons are going to be plotted. If they are widths and heights are made slightly larger}
}
\value{
Returns a list containing the heights and widths of the labels provided
}
\description{
Function used by \code{addTextLabels()}
}
\keyword{internal}
