% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextLabels.R
\name{chooseNewLocation}
\alias{chooseNewLocation}
\title{A function to choose (from the alternative locations) a new location for a label to be plotted at}
\usage{
chooseNewLocation(x, y, altXs, altYs)
}
\arguments{
\item{x}{The X coordinate at which label was to be plotted}

\item{y}{The Y coordinate at which label was to be plotted}

\item{altXs}{A vector of X coordinates for alternative locations}

\item{altYs}{A vector of Y coordinates for alternative locations}
}
\value{
Returns the index of the chosen alternative location
}
\description{
Function used by \code{addTextLabels()}
}
\keyword{internal}
