% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextLabels.R
\name{euclideanDistances}
\alias{euclideanDistances}
\title{Calculate the euclidean distance between two sets of points}
\usage{
euclideanDistances(x1s, y1s, x2s, y2s)
}
\arguments{
\item{x1s}{A vector X coordinates for the first set of points}

\item{y1s}{A vector Y coordinates for the first set of points}

\item{x2s}{A vector X coordinates for the second set of points}

\item{y2s}{A vector Y coordinates for the second set of points}
}
\value{
Returns the distances between the sets of points provided
}
\description{
Function used by \code{addTextLabels()}
}
\keyword{internal}
