% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextLabels.R
\name{tooClose}
\alias{tooClose}
\title{Checks whether a point is too close to any others}
\usage{
tooClose(xCoords, yCoords, index, textHeights, textWidths)
}
\arguments{
\item{xCoords}{A vector containing the X coordinates for labels}

\item{yCoords}{A vector containing the Y coordinates for labels}

\item{index}{The index of the point of interest}

\item{textHeights}{The heights of the labels to be plotted}

\item{textWidths}{The widths of the labels to be plotted}
}
\value{
Returns a logical variable to indicate whether the point of interest was too close to any points
}
\description{
Function used by \code{addTextLabels()}
}
\keyword{internal}
