% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextLabels.R
\name{chooseNewLocation}
\alias{chooseNewLocation}
\title{A function to choose (from the alternative locations) a new location for a label to be plotted at}
\usage{
chooseNewLocation(xCoords, yCoords, index, altXs, altYs, distances,
  textHeights, textWidths)
}
\arguments{
\item{xCoords}{The X coordinates of actual points}

\item{yCoords}{The Y coordinates of actual points}

\item{index}{The index of the point of interest}

\item{altXs}{A vector of X coordinates for alternative locations}

\item{altYs}{A vector of Y coordinates for alternative locations}

\item{distances}{The distances between the actual and alternative locations}

\item{textHeights}{The heights of the labels to be plotted}

\item{textWidths}{The widths of the labels to be plotted}
}
\value{
Returns the index of the chosen alternative location
}
\description{
Function used by \code{addTextLabels()}
}
\keyword{internal}
